requirements.txt
absl-py==2.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
asttokens==3.0.0
astunparse==1.6.3
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.3
contractions==0.1.73
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
emoji==2.15.0
en_core_web_sm 
executing==2.2.0
fastapi==0.121.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gensim==4.4.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.12.0
keras-tuner==1.4.7
kiwisolver==1.4.9
kt-legacy==1.0.5
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.3.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
murmurhash==1.0.13
mysql-connector-python==9.5.0
namex==0.1.0
narwhals==2.10.1
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.4
pillow==12.0.0
platformdirs==4.3.8
plotly==6.3.1
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.33.0
psutil==7.0.0
pure_eval==0.2.3
pyahocorasick==2.2.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.4.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.4.3
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
starlette==0.49.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
textsearch==0.0.24
thinc==8.3.6
threadpoolctl==3.6.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==2.0.0
wordcloud==1.9.4